Workaround performance issue on PLX import from many derivationParameters.
DerivedResponses with many derivationParameters take a very long time
to insert (bugzid:1427). To workaround, we store the entire .exp
contents as a *single* parameter. This works around slow performance on
ImportPLX, and lets us migrate to individual parameters in the future.
For now, querying by RegEx matching is still possible